embryoscope/requirements.txt
# Python 3.9+ required
pandas>=1.5.0
requests>=2.28.0
pyyaml>=6.0
duckdb>=0.9.0
urllib3>=1.26.0
tqdm>=4.64.0
psutil>=5.9.0 